requirements.txt
# For LLM interaction
google-generativeai
requests

# For fuzzy matching of product names
thefuzz
python-Levenshtein